set up pi db, and cleaned up models

+++ src/main/java/com/budgetapp/budgetapi/repo/BankAccountRepo.java
@@ -1,16 +1,15 @@
 package com.budgetapp.budgetapi.repo;
 
 import com.budgetapp.budgetapi.model.transaction.BankAccountModel;
-import com.budgetapp.budgetapi.model.transaction.BucketModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface BankAccountRepo extends JpaRepository<BankAccountModel, Long> {
 
-    BankAccountModel findByNameAndUserId(String name, Integer userId);
+    BankAccountModel findByNameAndUserId(String name, Long userId);
 
-    List<BankAccountModel> findAllByUserId(Integer userId);
+    List<BankAccountModel> findAllByUserId(Long userId);
 
-    BankAccountModel findByIdAndUserId(long bankAccountId, Integer id);
+    BankAccountModel findByIdAndUserId(Long bankAccountId, Long id);
 }
