+++ src/main/java/com/budgetapp/budgetapi/repo/BucketRepo.java
@@ -1,20 +1,18 @@
 package com.budgetapp.budgetapi.repo;
 
 import com.budgetapp.budgetapi.model.transaction.BucketModel;
-import com.budgetapp.budgetapi.model.transaction.TransactionModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Arrays;
 import java.util.List;
 
 public interface BucketRepo extends JpaRepository<BucketModel, Long> {
 
-    BucketModel findByNameAndUserId(String name, Integer userId);
+    BucketModel findByNameAndUserId(String name, Long userId);
 
 
-    List<BucketModel> findAllByUserId(Integer userId);
+    List<BucketModel> findAllByUserId(Long userId);
 
-    List<BucketModel> findAllByBankAccountIdAndUserId(Long accountId, Integer userId);
+    List<BucketModel> findAllByBankAccountIdAndUserId(Long accountId, Long userId);
 
-    BucketModel findByIdAndUserId(long l, Integer id);
+    BucketModel findByIdAndUserId(Long l, Long id);
 }
